+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Exchange view will auto refresh data [#1399](https://github.com/cloudamqp/lavinmq/pull/1399)
+
 ## [2.5.0-rc.4] - 2025-10-27
 
 ### Fixed